Convert new_transaction blueprint into module
Break into parts: forms, views
