+++ api_finances/api_finances/blueprints/new_transaction/views.py
@@ -3,25 +3,16 @@ from jinja2 import TemplateNotFound
 import wtforms
 from api_finances.utils.wtforms import ComboBoxField
 from api_finances import models
+from api_finances.blueprints.new_transaction import forms
 from werkzeug.datastructures import ImmutableMultiDict
 
 new_transaction = Blueprint('simple_page', __name__,
                         template_folder='templates')
 
 
-class NewTransactionForm(wtforms.Form):
-    date = wtforms.fields.DateField('Date', format='%Y-%m-%d')
-    total = wtforms.IntegerField('Total')
-    tax = wtforms.IntegerField('Tax')
-    payment_type = ComboBoxField('Payment Type')
-    retailer = ComboBoxField('Retailer')
-    description = wtforms.TextAreaField('Description')
-    #TODO: Add submit button and clear button
-
-
 @new_transaction.route('/', methods=['GET', 'POST'])
 def init():
-    form = NewTransactionForm(request.form)
+    form = forms.NewTransactionForm(request.form)
     retailer_names = [retailer.name for retailer in models.Retailer.query.all()]
     form.retailer.widget.options = retailer_names
     payment_type_names = [type.name for type in models.PaymentType.query.all()]
@@ -41,14 +32,14 @@ def init():
 
 @new_transaction.route('/update', methods=['GET', 'POST'])
 def add_transaction_update():
-    form = NewTransactionForm(request.form)
+    form = forms.NewTransactionForm(request.form)
     retailer_names = [retailer.name for retailer in models.Retailer.query.all()]
     form.retailer.widget.options = retailer_names
     payment_type_names = [type.name for type in models.PaymentType.query.all()]
     form.payment_type.widget.options = payment_type_names
     new_transaction_data = session.get('new_transaction_data', None)
     if request.method == 'GET' and new_transaction_data is not None:
-        form = NewTransactionForm(data=new_transaction_data)
+        form = forms.NewTransactionForm(data=new_transaction_data)
         form.payment_type.data = 'foobar'
         # form.date.data = new_transaction_data['date']
         # form.total.data = new_transaction_data['total']
@@ -71,10 +62,6 @@ def add_transaction_update():
         new_transaction_data=ImmutableMultiDict(new_transaction_data),
         form=form)
 
-class ConfirmTransactionForm(wtforms.Form):
-    confirm = wtforms.SubmitField('Confirm')
-    cancel = wtforms.SubmitField('Cancel')
-
 
 @new_transaction.route('/confirm', methods=['GET', 'POST'])
 def add_transaction_confirm():
@@ -82,7 +69,7 @@ def add_transaction_confirm():
     new_transaction_data = session.get('new_transaction_data', None)
     if new_transaction_data is None:
         return redirect(url_for('add_transaction_new'))
-    form = ConfirmTransactionForm(request.form)
+    form = forms.ConfirmTransactionForm(request.form)
     if request.method == 'POST' and form.validate():
         if form.confirm:
             return 'Confirm'
